[level 1] Title: 문자열을 정수로 바꾸기, Time: 0.02 ms, Memory: 9.32 MB -BaekjoonHub

+++ 프로그래머스/1/12925. 문자열을 정수로 바꾸기/문자열을 정수로 바꾸기.py
@@ -1,2 +1,2 @@
 def solution(s):
-    return float(s)
+    return int(s)
